// src/app/models/qastatus.ts
export interface Qastatus {
    qaBill_ID: number;
    qaDocNumber: string;
    userID: number;
    billDate: string;
    OrderNumber: string;
    saleManager: number;
    TotalQty: number;
    TotalFinishQty: number;
    TotalRepairQty: number;
    Balance: number;
    cusStatus: number;
    avatar: string;
    ProductionTeam: string;
    repairDate: Date;
    showIcon: string;
    TotalRepairFinish: number;
    nameReceive: string;
    thaiDate: string;
    nickNameReceive: string;

}