// src/app/models/qaReceiveMaster.ts
export interface QaReceiveMaster {
    qaBill_ID: number;
    qaDocNumber: string;
    userID: number;
    billDate: string;
    OrderNumber: string;
    salesManager: number;
    TotalQty: number;
    avatar: string;
    ProductionTeam: string;
    QtyOrder: number;
    DueDate: Date;
    TotalQaFinish: number;
    OrderQty: number;
    showIcon: string;
    userSent: string;
    nickNameSent: string;
    thaiDateBill: string;
    thaiDateDue: string;
}