// src/app/models/qaReceiveDetail.ts
export interface QaReceiveDetail {
    qaBill_ID: number;
    qaDocNumber: string;
    userID: number;
    billDate: string;
    receiver: number;
    receiveDate: string;
    qaBillStatus: string;
    bill_Item: string;
    OrderNumber: string;
    OrderItemNo: string;
    ProductID: string;
    Qty: number;
    ProductCode: string;
    ProductDesc: string;
    NewPict: string;
}