// src/app/models/qaFinishMaster.ts
export interface QaFinishMaster {

    qaFinish_ID: number;
    finishDocNumber: string;
    qaBill_ID: number;
    OrderNumber: string;
    fullName: string;
    receiveDate: Date;
    salesManager: number;
    dateLate: number;
    TotalQaFinish: number;
    showIcon: string;
    TotalQa: number;
    qaDocNumber: string;
    thaiDate: string;
    avatar: string;
    ProductionTeam: string;
    nickName: string;
}