// src/app/models/qaFinishDetail.ts
export interface QaFinishDetail {

    qaFinish_ID: number;
    finishDocNumber: string;
    qaBill_ID: number;
    userID: number;
    finishDate: Date;
    OrderNumber: string;
    ProductCode: string;
    NewPict: string;
    TotalQtyBill: number;
    TotalQtyFinish: number;
    finish_Item: number;

}