// src/app/models/orderStatusMaster.ts
export interface OrderStatusMaster {
    Balance: number;
    DueDate: Date;
    OrderDate: Date;
    OrderNumber: string;
    ProductionTeam: string;
    SalesManager: string;
    TotalQaBill: number;
    TotalQaFinish: number;
    TotalQaRepair: number;
    TotalQty: number;
    Totalrepairsum: number;
    avatar: string;
    dateLate: number;
    saleId: number;
    thaiDate: string;
    JobDay: number;
}